refactor(backup): camelCase contributor files
Rename the kebab-case backupContributor-*.ts files to camelCase to comply
with naming-conventions §3 (src/main files must be camelCase or PascalCase).
The kebab names tripped the naming/path-case lint rule.

27 files renamed (13 contributor modules + 14 test files) across
src/main/data and src/main/data/services; all import paths updated
(CONTRIBUTORS barrel + cross-test imports). Pure rename, no behavior change.
197 contributor/merge/finalize tests green; typecheck + eslint naming clean.

Note: backupContributor-translate.test.ts is a misnamed orphan (it tests
@main/services/translate/backupContributor, not a backupContributor-translate
module); renamed to camelCase here to clear the lint rule, but its location
and semantics warrant a separate follow-up.
